requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
graphql-core==3.1.7
greenlet==1.1.2
h11==0.12.0
httptools==0.2.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
orjson==3.6.5
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
pydantic==1.8.2
Pygments==2.10.0
pyparsing==3.0.6
pytest==6.2.5
pytest-sugar==0.9.4
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
sentinel==0.3.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.28
starlette==0.16.0
strawberry-graphql==0.93.10
termcolor==1.1.0
toml==0.10.2
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
